Like and Unlike

+++ server/controllers/post/deletePost.ts
@@ -1,7 +1,8 @@
-import { Request, Response, RequestHandler, NextFunction } from "express";
+import { Response, RequestHandler, NextFunction } from "express";
 import asyncHandler from "express-async-handler";
 import Post from "../../models/Post";
 import CustomErrorHandler from "../../services/CustomErrorHandler";
+import IAuthUserRequest from "../../interfaces/AuthUser";
 
 /**
  * @description   GET user posts
@@ -10,8 +11,14 @@ import CustomErrorHandler from "../../services/CustomErrorHandler";
  */
 
 const deletePost: RequestHandler = asyncHandler(
-  async (req: Request, res: Response, next: NextFunction) => {
-    // Get posts
+  async (req: IAuthUserRequest, res: Response, next: NextFunction) => {
+    const userId = req.user?._id;
+
+    if (userId?.toString() !== req.params.userId) {
+      return next(CustomErrorHandler.unAuthorized("You can not create post!"));
+    }
+
+    // Delete posts
     try {
       const post = await Post.findByIdAndDelete({ _id: req.params.postId });
       res.status(202).json(post);
